requirements.txt
pydantic~=1.8.2
fastapi~=0.70.0
uvicorn[standard]
typer~=0.4.0
tortoise-orm~=0.17.8
starlette~=0.16.0
requests~=2.26.0